Backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==6.0
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyPDF2==3.0.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
urllib3==2.5.0
